Move ListWatch to executor pool manager
