+++ controller/controller.go
@@ -17,7 +17,6 @@ limitations under the License.
 package controller
 
 import (
-	"context"
 	"log"
 
 	"github.com/fission/fission"
@@ -40,10 +39,7 @@ func Start(port int) {
 
 	fc.WaitForCRDs()
 
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	api, err := MakeAPI(ctx)
+	api, err := MakeAPI()
 	if err != nil {
 		log.Fatalf("Failed to start controller: %v", err)
 	}
